SCHRAMM-47 fix (defaults,tasks): Rename some variables to the correct naming standard
All variables in a role should be named as follows:

${ROLE_NAME}_${VARIABLE_NAME}

This is to make it clear to what role the variables apply in a
group_vars.yml document, as well as as to remind the developer that
roles are independent and variables defined for one role should not be
used in another.

This role was using two roles incorrectly:

- mysql_database_name
- mysql_database_name

There variables were defined for another role originally, and consumed
in this role. However, as this role is now being used without that other
role, this role is broken.

This commit renames the variables to the correct standard so they can be
implemented in a backwards compatible way.

BREAKING CHANGE: This renames some variables. You will need to
re-implement these in your playbook.