AD-HOC refactor (AWS Keys): Lookup keys from variables
At this time, AWS keys are being looked up out of the process
environment. However, there are some cases (such as in the provisioning
of a local machine with local ansible) that the process environment will
not contain this authorisation data, and that it will instead be passed
by the --extra-vars Ansible configuration.

This commit replaces the existing direct lookup of the environment with
that variables, but sources the content of those variables from the
environment. In this way, the existing behaviour is preserved, but with
the ability to extend it in the specific cases required.

Future work might seek to remove the environment lookup entirely from
this playbook, however that is a breaking change, and thus should be
subject to a deprecation period and a documented migration plan --
something that does not seem a reasonable solution at this time.