AD-HOC refactor (Dependencies): Use pip to manage dependencies
When this role was initially created, it was presumed that it will
always be be used with Python 3, and managed accordingly. However, with
further usage, it is being used with python 2. This, the package manager
for python packages was changed to "pip". This will correctly resolve
dependencies regardless of the interpreter.

Design Notes:

own dependencies:

  More generally, the author believes that roles should be as portable
  and effortless to reuse as possible. This invariably means managing
  their own dependencies.

separate dependencies task:

  More recently the author has been splitting the tasks out by
  responsibility. "Dependencies" is one obvious split from the
  canonical purpose of this role.